手動入力機能をhooks/useBookScannerに統合し、不要なBarcodeScannerコンポーネントを削除
