+++ hooks/useBookScanner.ts
@@ -1,6 +1,11 @@
 import { useState, useRef, useCallback } from "react";
 import { Alert } from "react-native";
 import { RakutenBookService } from "@/services/RakutenBookService";
+import { BookStorageService } from "@/services/BookStorageService";
+import { Book } from "@/constants/MockData";
+
+// No-Image用のデフォルト画像URL
+const DEFAULT_BOOK_IMAGE = "https://via.placeholder.com/150x200?text=No+Image";
 
 interface UseBookScannerProps {
   onClose: () => void;
@@ -10,9 +15,64 @@ export const useBookScanner = ({ onClose }: UseBookScannerProps) => {
   const [processedISBNs, setProcessedISBNs] = useState<Set<string>>(new Set());
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [isAlertShowing, setIsAlertShowing] = useState<boolean>(false);
+  // 手動入力フォーム用の状態
+  const [showManualEntryForm, setShowManualEntryForm] =
+    useState<boolean>(false);
+  const [bookTitle, setBookTitle] = useState<string>("");
+  const [bookAuthor, setBookAuthor] = useState<string>("");
+  const [scannedBarcode, setScannedBarcode] = useState<string>("");
+
   const isProcessingRef = useRef(false);
   const lastProcessedTime = useRef(0);
   const DEBOUNCE_TIME = 1000; // 1秒のデバウンス時間
+
+  // 手動入力フォームを表示する関数
+  const showManualForm = useCallback(() => {
+    setShowManualEntryForm(true);
+  }, []);
+
+  // 手動入力フォームを閉じる関数
+  const hideManualForm = useCallback(() => {
+    setShowManualEntryForm(false);
+    setBookTitle("");
+    setBookAuthor("");
+  }, []);
+
+  // 手動入力された書籍情報を保存する関数
+  const handleManualSave = useCallback(async () => {
+    if (!bookTitle.trim()) {
+      Alert.alert("エラー", "書籍名は必須です");
+      return;
+    }
+
+    try {
+      // 手動で入力された情報から書籍オブジェクトを作成
+      const newBook: Book = {
+        id: `manual_${Date.now()}`,
+        title: bookTitle.trim(),
+        author: bookAuthor.trim() || "不明",
+        coverImage: DEFAULT_BOOK_IMAGE,
+      };
+
+      // 書籍を保存
+      await BookStorageService.saveBook(newBook);
+
+      Alert.alert("保存完了", `「${newBook.title}」を本棚に追加しました。`, [
+        {
+          text: "OK",
+          onPress: () => {
+            hideManualForm();
+            onClose();
+          },
+        },
+      ]);
+    } catch (error) {
+      console.error("Error saving manual book entry:", error);
+      Alert.alert("エラー", "書籍の保存中にエラーが発生しました。", [
+        { text: "OK" },
+      ]);
+    }
+  }, [bookTitle, bookAuthor, hideManualForm, onClose]);
 
   const handleBarcodeScanned = useCallback(
     async (isbn: string) => {
@@ -30,9 +90,9 @@ export const useBookScanner = ({ onClose }: UseBookScannerProps) => {
       console.log("ISBN検出:", isbn);
       isProcessingRef.current = true;
       setProcessedISBNs((prev) => new Set(prev).add(isbn));
+      setScannedBarcode(isbn); // 検出したISBNを保存
       setIsLoading(true);
       setIsAlertShowing(true);
-      setIsLoading(true);
 
       Alert.alert(
         "ISBN検出",
@@ -56,13 +116,23 @@ export const useBookScanner = ({ onClose }: UseBookScannerProps) => {
                 if (!result) {
                   Alert.alert(
                     "書籍が見つかりません",
-                    `ISBN ${isbn} に一致する書籍が見つかりませんでした。`,
+                    `ISBN ${isbn} に一致する書籍が見つかりませんでした。手動で入力しますか？`,
                     [
                       {
-                        text: "OK",
+                        text: "いいえ",
+                        style: "cancel",
                         onPress: () => {
                           setIsAlertShowing(false);
                           isProcessingRef.current = false;
+                        },
+                      },
+                      {
+                        text: "はい",
+                        onPress: () => {
+                          setIsAlertShowing(false);
+                          isProcessingRef.current = false;
+                          setIsLoading(false);
+                          showManualForm();
                         },
                       },
                     ]
@@ -155,13 +225,17 @@ export const useBookScanner = ({ onClose }: UseBookScannerProps) => {
         ]
       );
     },
-    [isAlertShowing, processedISBNs, onClose]
+    [isAlertShowing, processedISBNs, onClose, showManualForm]
   );
 
   const resetScanner = useCallback(() => {
     setProcessedISBNs(new Set());
     setIsLoading(false);
     setIsAlertShowing(false);
+    setShowManualEntryForm(false);
+    setBookTitle("");
+    setBookAuthor("");
+    setScannedBarcode("");
     isProcessingRef.current = false;
     lastProcessedTime.current = 0;
   }, []);
@@ -171,5 +245,14 @@ export const useBookScanner = ({ onClose }: UseBookScannerProps) => {
       isLoading || isAlertShowing ? () => {} : handleBarcodeScanned,
     isLoading,
     resetScanner,
+    // 手動入力関連の値を追加
+    showManualEntryForm,
+    showManualForm,
+    hideManualForm,
+    bookTitle,
+    setBookTitle,
+    bookAuthor,
+    setBookAuthor,
+    handleManualSave,
   };
 };
